nearest point

+++ MemeLib/MemeLib-Game/FaceSteering.h
@@ -3,7 +3,7 @@
 class FaceSteering :public Steering
 {
 private:
-	double mTimeToTarget;
+	float mTimeToTarget;
 	float mTargetRadius, mSlowRadius;
 public:
 	FaceSteering(const UnitID& ownerID, const Vec3& targetLoc, const UnitID& targetID, float theRadius, float theSlowRadius);
